Separate validation_util into integration_test_util
